+++ lib/simple_main_screen.dart
@@ -368,7 +368,7 @@ class _SimpleMainScreenState extends State<SimpleMainScreen> {
             _buildServerCard(),
             const SizedBox(height: 32),
             Text(
-              'Simple Degoogled Traccar Client v9.5.2d\nNo Google Services Required',
+              'Simple Degoogled Traccar Client v9.5.2e\nNo Google Services Required',
               textAlign: TextAlign.center,
               style: Theme.of(context).textTheme.bodySmall,
             ),